Renamed the folder from "model" to "src" move some files
