+++ monocliche/src/Board.py
@@ -1,6 +1,6 @@
 from typing import Optional
 
-from monocliche import Constants
+from monocliche.src import Constants
 
 from monocliche.model.Box import Box
 from monocliche.model.Deck import Deck
