+++ monocliche/src/DiceRollResult.py
