+++ monocliche/src/Region.py
@@ -1,4 +1,4 @@
-from monocliche import Constants
+from monocliche.src import Constants
 
 from monocliche.model.Property import Property
 from monocliche.model import DiceRollResult
