+++ monocliche/tests/DiceRollResultTest.py
